requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
et-xmlfile==1.1.0
fastapi==0.112.1
fonttools==4.53.1
frozenlist==1.4.1
gotrue==2.7.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.8
kiwisolver==1.4.5
lxml==5.3.0
matplotlib==3.9.2
multidict==6.0.5
numpy==1.26.0
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
postgrest==0.16.11
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
pyshorteners==1.0.1
pytest-mock==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==1.0.2
realtime==2.0.2
reportlab==3.6.13
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.3
supafunc==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
XlsxWriter==3.2.0
yarl==1.9.4
